rename store provider
